+++ platform/android/MapLibreAndroid/src/main/java/org/maplibre/android/style/sources/RasterSource.kt
@@ -45,6 +45,10 @@ class RasterSource : Source {
      *
      *  * load resources from the Android file system
      *
+     *  * file:///android_asset/
+     *
+     *  * load resources from the APK assets folder (same as asset://)
+     *
      *  * asset://
      *
      *  * load resources from the binary packaged assets folder
@@ -71,6 +75,10 @@ class RasterSource : Source {
      *  * file://
      *
      *  * load resources from the Android file system
+     *
+     *  * file:///android_asset/
+     *
+     *  * load resources from the APK assets folder (same as asset://)
      *
      *  * asset://
      *
@@ -100,6 +108,10 @@ class RasterSource : Source {
      *  * file://
      *
      *  * load resources from the Android file system
+     *
+     *  * file:///android_asset/
+     *
+     *  * load resources from the APK assets folder (same as asset://)
      *
      *  * asset://
      *