+++ platform/android/MapLibreAndroid/src/main/java/org/maplibre/android/style/sources/VectorSource.kt
@@ -53,6 +53,10 @@ class VectorSource : Source {
      *
      *  * load resources from the Android file system
      *
+     *  * file:///android_asset/
+     *
+     *  * load resources from the APK assets folder (same as asset://)
+     *
      *  * asset://
      *
      *  * load resources from the binary packaged assets folder
@@ -79,6 +83,10 @@ class VectorSource : Source {
      *  * file://
      *
      *  * load resources from the Android file system
+     *
+     *  * file:///android_asset/
+     *
+     *  * load resources from the APK assets folder (same as asset://)
      *
      *  * asset://
      *